W-073 - move entities to other package
