+++ backend/plortal/src/main/java/com/plortal/plortal/controller/ScheduleNotesController.java
@@ -1,6 +1,6 @@
 package com.plortal.plortal.controller;
 
-import com.plortal.plortal.model.ScheduleNotes;
+import com.plortal.plortal.model.entity.ScheduleNotes;
 import com.plortal.plortal.service.ScheduleNotesService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
