+++ backend/plortal/src/main/java/com/plortal/plortal/model/entity/Location.java
@@ -1,4 +1,4 @@
-package com.plortal.plortal.model;
+package com.plortal.plortal.model.entity;
 
 import com.plortal.plortal.model.enums.LocationType;
 import jakarta.persistence.*;
