+++ backend/plortal/src/main/java/com/plortal/plortal/repository/NewsRepository.java
@@ -1,6 +1,6 @@
 package com.plortal.plortal.repository;
 
-import com.plortal.plortal.model.News;
+import com.plortal.plortal.model.entity.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsRepository extends JpaRepository<News, Long> {
