+++ backend/plortal/src/main/java/com/plortal/plortal/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.plortal.plortal.repository;
 
-import com.plortal.plortal.model.User;
+import com.plortal.plortal.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
