+++ backend/plortal/src/main/java/com/plortal/plortal/service/LocationService.java
@@ -1,6 +1,6 @@
 package com.plortal.plortal.service;
 
-import com.plortal.plortal.model.Location;
+import com.plortal.plortal.model.entity.Location;
 import com.plortal.plortal.model.enums.LocationType;
 
 import java.util.List;
