+++ backend/plortal/src/main/java/com/plortal/plortal/service/impl/ScheduleNotesServiceImpl.java
@@ -1,6 +1,6 @@
 package com.plortal.plortal.service.impl;
 
-import com.plortal.plortal.model.ScheduleNotes;
+import com.plortal.plortal.model.entity.ScheduleNotes;
 import com.plortal.plortal.repository.ScheduleNotesRepository;
 import com.plortal.plortal.service.ScheduleNotesService;
 import org.springframework.beans.factory.annotation.Autowired;
