+++ backend/plortal/src/main/java/com/plortal/plortal/service/impl/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.plortal.plortal.service.impl;
 
 import com.plortal.plortal.exception.*;
-import com.plortal.plortal.model.User;
+import com.plortal.plortal.model.entity.User;
 import com.plortal.plortal.model.enums.UserRole;
 import com.plortal.plortal.repository.UserRepository;
 import com.plortal.plortal.service.UserService;
